#704 finite volumes tests all pass
